optimize C code: using int for calc distance

+++ geo-city/harvesine/harvesine.pyx
@@ -4,7 +4,7 @@ from libc.stdlib cimport free
 
 cdef extern from "module.c" nogil:
     double harvesine_distance_impl(double lon_1, double lat_1, double lon_2, double lat_2);
-    double* combinations_impl(float *queue_values, float *elemts_values, int elemts_size);
+    int* combinations_impl(float *queue_values, float *elemts_values, int elemts_size);
 
 cdef double harvesine_distance(double lon_1, double lat_1, double lon_2, double lat_2) nogil:
     return harvesine_distance_impl(lon_1, lat_1, lon_2, lat_2);
@@ -15,18 +15,18 @@ def calculate(double lon_1, double lat_1, double lon_2, double lat_2):
         result = harvesine_distance(lon_1, lat_1, lon_2, lat_2);
     return result
 
-cdef convert_to_python(double *ptr, int n):
+cdef convert_to_python(int *ptr, int n):
     cdef int i
     lst=[]
     for i in range(n):
         lst.append(ptr[i])
     return lst
 
-cdef double* harvesine_array(float *queue_values, float *elemts_values, int elemts_size) nogil:
+cdef int* harvesine_array(float *queue_values, float *elemts_values, int elemts_size) nogil:
     return combinations_impl(queue_values, elemts_values, elemts_size);
 
 def combinations(queue_values, elemts_values, elemts_size):
-    cdef double* result;
+    cdef int* result;
     cdef array.array queue_values_arr = array.array('f', queue_values);
     cdef array.array elemts_values_arr = array.array('f', elemts_values);
     result = harvesine_array(queue_values_arr.data.as_floats, elemts_values_arr.data.as_floats, elemts_size);
